Rename different CheckIcons to avoid confusion
