+++ src/icons/SwitchCheckIcon.vue
